+++ opendbc/car/tesla/carcontroller.py
@@ -2,7 +2,7 @@ import numpy as np
 import math
 from opendbc.can import CANPacker
 from opendbc.car import ACCELERATION_DUE_TO_GRAVITY, Bus, DT_CTRL, rate_limit
-from opendbc.car.lateral import AngleSteeringLimits
+from opendbc.car.controls.lateral import AngleSteeringLimits
 from opendbc.car.interfaces import CarControllerBase, ISO_LATERAL_ACCEL
 from opendbc.car.tesla.teslacan import TeslaCAN
 from opendbc.car.tesla.values import CarControllerParams
